requirements.txt
# Install these requirements if you wish to contribute to this project.

# ===========================================================================
# Packages essential to this app. Needed by anyone who wants to use this app.
# ===========================================================================
django==1.4.3
django-cms==2.3.5

# There is a problem with staticfiles in the current pypi release
# cmsplugin-blog==1.1.2
-e git+git://github.com/fivethreeo/cmsplugin-blog.git@cd91a4eb066437aa254574759eed51da973314ab#egg=cmsplugin_blog
django-filer==0.9.3
django-document-library==0.2
South==0.7.6


# ==============================================================
# Packages needed for running the tests. Needed by contributors.
# ==============================================================
django-libs==0.14.1
Fabric
mock
django-nose
coverage
django-coverage
-e git+git://github.com/jcrocholl/pep8.git#egg=pep8
-e git+git://github.com/bmcustodio/flake8.git#egg=flake8
ipython
ipdb
